vision/requirements.txt
opencv-python==4.8.1.78
ultralytics==8.0.200
paho-mqtt==1.6.1
pyyaml==6.0.1
requests==2.31.0
numpy==1.24.3
